Google Authentication - in progress, see notes
Pushing my update for now. Currently the google sign in alert pops up but then disappears. I'm getting a warning about the viewcontroller being deallocated which might cause "undefined behavior"

Most changes are in LogInView.swift, created signIn functions that should make a POST request to backend and pass in the token and get back a response.

I created an AppDelegate to simulate a UIKit environment and manage the lifecycle of the delegate. Not sure how to integrate that into my solution.

Historically google auth has worked with UIKit, but there's not much guidance anywhere for how to port it to SwiftUI. Even worse, GoogleSignIn went through a major update to v6 recently so any possible guidance online is most likely obsolete now.